requirements.txt
ema_pytorch==0.7.7
kaggle==1.7.4.2
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
Pillow==11.1.0
python-dotenv==1.1.0
Requests==2.32.3
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
wandb==0.19.11
opencv-python==4.11.0.86
torchinfo==1.8.0
fightingcv-attention==1.0.0
grad-cam==1.5.5
seaborn==0.13.2
scikit-learn==1.5.1
kagglehub==0.3.12
